Restructure: Move commands package folder up to src root
