+++ src/commands/new_training_poll_impl.go
